refactor: отделены сервисы для модели значения
